+++ src/arize/_generated/api_client/test/test_experiment_run.py
@@ -37,13 +37,13 @@ class TestExperimentRun(unittest.TestCase):
             return ExperimentRun(
                 id = '',
                 example_id = '',
-                output = ''
+                output = '',
+                error = ''
             )
         else:
             return ExperimentRun(
                 id = '',
                 example_id = '',
-                output = '',
         )
         """