+++ src/arize/prompts/types.py
@@ -8,6 +8,7 @@ from arize._generated.api_client.models.invocation_params import (
 )
 from arize._generated.api_client.models.llm_message import LLMMessage
 from arize._generated.api_client.models.llm_provider import LlmProvider
+from arize._generated.api_client.models.message_role import MessageRole
 from arize._generated.api_client.models.prompt import Prompt
 from arize._generated.api_client.models.prompt_version import PromptVersion
 from arize._generated.api_client.models.prompt_version_labels_set200_response import (
@@ -23,12 +24,14 @@ from arize._generated.api_client.models.prompts_list200_response import (
     PromptsList200Response,
 )
 from arize._generated.api_client.models.provider_params import ProviderParams
+from arize._generated.api_client.models.tool_call_type import ToolCallType
 
 __all__ = [
     "InputVariableFormat",
     "InvocationParams",
     "LLMMessage",
     "LlmProvider",
+    "MessageRole",
     "Prompt",
     "PromptVersion",
     "PromptVersionLabelsSet200Response",
@@ -36,4 +39,5 @@ __all__ = [
     "PromptWithVersion",
     "PromptsList200Response",
     "ProviderParams",
+    "ToolCallType",
 ]